requirements.txt
pytest==6.2.5
pytest-env==0.6.2
allure-pytest==2.12.0
selenium==4.8.3
webdriver_manager==3.8.5
PyYAML==6.0
